Adding in-memory kv and better kv tests
