+++ cmd/tarmac/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/madflojo/tarmac/app"
+	"github.com/madflojo/tarmac/pkg/app"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 	_ "github.com/spf13/viper/remote"
@@ -22,6 +22,8 @@ func main() {
 	cfg.SetDefault("config_watch_interval", 15)
 	cfg.SetDefault("wasm_function", "/functions/tarmac.wasm")
 	cfg.SetDefault("wasm_function_config", "/functions/tarmac.json")
+	cfg.SetDefault("kvstore_type", "internal")
+	cfg.SetDefault("data_dir", "/data/tarmac")
 	cfg.SetDefault("grpc_socket_path", "/grpc.sock")
 
 	// Load Config
